refactor: 필요없는 로직은 하위 컴포넌트로 옮기기
